% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_site}
\alias{build_site}
\alias{build_articles}
\alias{build_article}
\title{Website build helpers}
\usage{
build_site()

build_articles()

build_article(id = "sc")
}
\arguments{
\item{id}{character, the ID (\code{"sc"} or \code{"rr"}) for the Stellar Cartography and the Rest and Relaxation pages, respectively.}
}
\value{
nothing is returned but files on disk are edited.
}
\description{
Helper functions for building the \code{rtrek} package website.
}
\details{
These functions wrap around \code{pkgdown} and make html substitutions to customize the \code{rtrek} package website.
These functions should only be called when working on the \code{rtrek} package with that package root directory as the working directory.
}
\examples{
\dontrun{build_site()}
}
