% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\name{st_epub_test}
\alias{st_epub_test}
\title{Testing function for Star Trek books}
\usage{
st_epub_test(file, details = FALSE, add_tail = FALSE,
  default_reader = FALSE)
}
\arguments{
\item{file}{character.}

\item{details}{logical.}

\item{add_tail}{logical.}

\item{default_reader}{logical.}
}
\value{
returns invisibly
}
\description{
A testing function for reading Star Trek epub files with \code{epubr::epub} or \code{st_epub}.
}
\examples{
\dontrun{st_epub_test(file)}
}
