% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes.R
\name{st_script_download}
\alias{st_script_download}
\title{Download episode scripts}
\usage{
st_script_download(download_dir = tempdir(), keep = FALSE,
  overwrite = FALSE)
}
\arguments{
\item{download_dir}{download directory.}

\item{keep}{logical, if \code{FALSE} (default) then downloaded files are removed after processing.}

\item{overwrite}{logical, if \code{FALSE} (default) then no downloading occurs for any file already present in \code{download_dir}.}
}
\value{
a data frame; may write files to disk.
}
\description{
Download episode scripts into a nested tidy data frame.
}
\details{
Downloads almost entirely come from \code{scifi.media}, but when unavailable are pulled from \code{chakoteya.net}.
}
\examples{
\dontrun{x <- st_script_download("data-raw/episode_scripts", TRUE)}
}
