% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes.R
\name{st_script_info}
\alias{st_script_info}
\title{Episode Script Information}
\usage{
st_script_info()
}
\value{
a data frame
}
\description{
Star Trek episode and movie transcript metadata.
}
\examples{
\dontrun{st_script_info()}
}
