% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tame_tweets}
\alias{tame_tweets}
\title{Tame tweets}
\usage{
tame_tweets(x, accounts = "RikerGoogling")
}
\arguments{
\item{x}{data frame}

\item{accounts}{character, limit to these accounts.}
}
\value{
a data frame
}
\description{
Make tweet tame.
}
\details{
This functions removes potentially offensive tweets using a list of stop words.
}
\examples{
\dontrun{tame_tweets(x)}
}
