% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_entries.R
\name{tl_entries}
\alias{tl_entries}
\title{Timeline Entries Data Frame}
\usage{
tl_entries(x)
}
\arguments{
\item{x}{character.}
}
\value{
a data frame
}
\description{
Create the timeline entries data frame.
}
\examples{
\dontrun{tl_entries(x)}
}
