% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl.R
\name{tl_footnotes}
\alias{tl_footnotes}
\title{Timeline Footnotes Data Frame}
\usage{
tl_footnotes(x, ibid = FALSE)
}
\arguments{
\item{x}{character.}

\item{ibid}{logical, leave Ibid entries as is.}
}
\value{
a data frame
}
\description{
Create the timeline footnotes data frame.
}
\examples{
\dontrun{tl_footnotes(x)}
}
