% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_local_tweets}
\alias{update_local_tweets}
\title{Update tweets}
\usage{
update_local_tweets(users, file, n = 3200)
}
\arguments{
\item{users}{character.}

\item{file}{character.}

\item{n}{integer, maximum number of tweets to go back chronologically from the latest.}
}
\value{
a data frame, and side effect of appending to the data frame saved in \code{file}
}
\description{
Update tweets saved to disk.
}
\details{
This function loads existing tweets from disk, downloads newer tweets, appends to the data frame, and overwrites the original file with the appended data frame, and returns the full data frame.
}
\examples{
\dontrun{update_local_tweets("PicardTips", myfile)}
}
