% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\name{st_book_copy}
\alias{st_book_copy}
\title{Copy Star Trek books}
\usage{
st_book_copy(in_dir, out_dir)
}
\arguments{
\item{in_dir}{character.}

\item{out_dir}{character.}
}
\value{
returns invisibly
}
\description{
Copy Star Trek epub files, with some changes.
}
\examples{
\dontrun{st_book_copy(in_dir, out_dir)}
}
