% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\name{st_epub}
\alias{st_epub}
\title{Read Star Trek epub files}
\usage{
st_epub(file, fields = .st_fields, chapter_pattern = .st_pat,
  add_pattern = .st_pat_list(), cleaner = clean_book_text,
  drop_sections = .st_sec_drop, series = TRUE, fix_date = TRUE,
  fix_text = TRUE, dedication = TRUE, hist_note = TRUE)
}
\arguments{
\item{file}{character.}

\item{fields}{character.}

\item{chapter_pattern}{character.}

\item{add_pattern}{list.}

\item{cleaner}{function.}

\item{drop_sections}{character.}

\item{series}{logical.}

\item{fix_date}{logical.}

\item{fix_text}{logical.}

\item{dedication}{logical.}

\item{hist_note}{logical.}
}
\value{
a data frame
}
\description{
A wrapper around \code{epubr::epub} for Star Trek epub files.
}
\examples{
\dontrun{st_epub(file)}
}
