% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes.R
\name{st_script_text_df}
\alias{st_script_text_df}
\title{Curate nested scripts to data frames}
\usage{
st_script_text_df(x)
}
\arguments{
\item{x}{a vector of lines of script. See example.}
}
\value{
a data frame
}
\description{
Curate nested vectors of script lines to data frames.
}
\details{
This currently works well (but imperfectly) for most scripts; it does not yet work for ENT, TAS, or a couple TOS scripts.
}
\examples{
\dontrun{
x <- st_script_download("data-raw/episode_scripts", TRUE)
system.time(x <- dplyr::mutate(x, text2 = purrr::map(text, st_script_text_df))) # ~6 minutes
}
}
